% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{analyser}
\alias{analyser}
\title{Analyse an xts object}
\usage{
analyser(x, threshold, window)
}
\arguments{
\item{x}{The xts object to be analyzed.}

\item{threshold}{The minimum threshold.}

\item{window}{The search window.}
}
\value{
A 2-column matrix with start and end indices.
}
\description{
Get start and end points where a given threshold is
satisfied.
}
