% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_detection.R
\name{event_detection}
\alias{event_detection}
\title{Detect rainfall events from given timereries. values are interpreted as mm/dt}
\usage{
event_detection(x, value_threshold = 0.01, event_threshold = 0.5,
  n_max_event_time = 48, n_roll_max_event = 12)
}
\arguments{
\item{x}{The xts object to be analyzed.}

\item{value_threshold}{The minimum threshold for individual timestep.}

\item{event_threshold}{The threshold for relevant event within sep-time}

\item{n_max_event_time}{The search window / separation time per event.}

\item{n_roll_max_event}{# timesteps for max sum over n}
}
\value{
tibble with event details
}
\description{
Detect rainfall events from given timereries. values are interpreted as mm/dt
}
