% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rainfall_events}
\alias{rainfall_events}
\title{Analyse an xts object}
\usage{
rainfall_events(x, value_threshold = 0.01, event_threshold = 0.5,
  n_max_event_time = 48L, n_roll_max_event = 12L)
}
\arguments{
\item{x}{The xts object to be analyzed.}

\item{value_threshold}{The minimum threshold for individual timestep.}

\item{event_threshold}{The threshold for relevant event within sep-time}

\item{n_max_event_time}{The search window / separation time per event.}
}
\value{
A 2-column matrix with start and end indices.
}
\description{
Get start and end points where a given threshold and event sum 
within a given time is satisfied.
}
