% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{split_by_ep}
\alias{split_by_ep}
\title{Split an xts-object into a list of xts-objects by eventpoints}
\usage{
split_by_ep(x, m)
}
\arguments{
\item{x}{An xts-object}

\item{m}{A two-column matrix containing eventpoints}
}
\value{
A list of xts-objects
}
\description{
Split an xts-object into a list of xts-objects by eventpoints
}
